+++ doc/changelog.md
@@ -1,2 +1,6 @@
 ## RC-0.2.0
-Renamed RelayProof to Proof
+- Renamed RelayProof to Proof
+- Renamed Invoice (memory) to Evidence
+- Renamed StoredInvoice (blockchain persisted) to Receipt
+- Renamed ProofWaitingPeriod to ClaimSubmissionWindow
+- Changed RPC and from `node-proof` to `node-receipt`